% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/afcp.R
\name{make.wide.data}
\alias{make.wide.data}
\title{Make the dataset wide from the default long to estimate AFCPs - drops all observations w/o level_a or level_b}
\usage{
make.wide.data(
  indata,
  attr_var,
  level_a,
  level_b,
  respondentID,
  choice,
  qid,
  option
)
}
\arguments{
\item{indata}{Dataframe containing the "long" dataset from \code{cjoint()}}

\item{attr_var}{Character name of the attribute to split on}

\item{level_a}{Character denoting the name of the first attribute level of interest}

\item{level_b}{Character denoting the name of the second attribute level of interest}

\item{respondentID}{Character denoting the column identifying the unique respondent}

\item{choice}{Character denoting the column identifying the binary choice outcome variable}

\item{qid}{Character denoting the question identifier}

\item{option}{Character denoting the identifier for each option/choice}
}
\value{
A dataframe with the following columns:
\itemize{
\item \code{respid} - Respondent identifier
\item \code{qid} - Task/question identifier
\item \code{choose} - An indicator for whether the profile with \code{val1} is selected in this task.
\item \code{treatment} - Character concatenating the two levels of interest \code{val1} and \code{val2}
\item \code{val1} - Level assigned to the attribute of interest for profile 1
\item \code{val2} - Level assigned to the attribute of interest for profile 2
}
}
\description{
Make the dataset wide from the default long to estimate AFCPs - drops all observations w/o level_a or level_b
}
