% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/afcp.R
\name{afcp}
\alias{afcp}
\title{Estimates AFCPs from a cjoint object}
\usage{
afcp(
  cjointobj,
  respondent.id,
  task.id,
  profile.id,
  attribute,
  baseline = NULL,
  ci = 0.95
)
}
\arguments{
\item{cjointobj}{Object of class \code{amce} returned by the \code{amce()} function in the \code{cjoint} package}

\item{respondent.id}{Character denoting the column identifying the unique respondent in the dataset from \code{cjointobj}}

\item{task.id}{Character denoting the column identifying the task/question in the dataset from \code{cjointobj}}

\item{profile.id}{Character denoting the column identifying the profile in each question in the dataset from \code{cjointobj} - this column should take on either 1 or 2 for a binary choice task}

\item{attribute}{Character denoting the attribute of interest}

\item{baseline}{Character denoting the level of the attribute to be used as the baseline. Default of \code{NULL} will pull the baseline from \code{cjointobj}.}

\item{ci}{Numeric between 0 and 1 denoting the size of the confidence interval to report for each afcp}
}
\value{
A list containing three dataframes
\itemize{
\item \code{afcp} - A dataframe containing estimated AFCPs relative to the selected baseline level.
\item \code{wald} - A dataframe containing the results for the aggregate Wald test of the equivalence of direct and indirect preferences across all other levels
\item \code{wald_three_level} - A dataframe containing the direct (AFCP) and indirect (difference in AFCPs) preference estimates across all other levels
}
}
\description{
This estimates the Average Feature Choice Probability (AFCP) for a given attribute and level combination
using the output from an amce object returned by the cjoint package
}
